chore: adding use hooks on search
